migrate from REAL to double

+++ FEM/PDEs/Node.hpp
@@ -8,7 +8,7 @@ class Node
 	bool _isBoundary;
 public:
 	Node(unsigned, Point, bool isBoundary = false);
-	Node(unsigned, REAL, REAL, REAL, bool isBoundary = false);
+	Node(unsigned, double, double, double, bool isBoundary = false);
 	void setLocation(Point const &);
 	void setIndex(unsigned);
 	void setAsBoundary(bool value = true);
